requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.1
python-multipart==0.0.6
